refactor(deps): relocate deriveServerID to internal/serverid (decouple PR-B' slice A)
First slice of relocating the monitor/supervisor's remaining package-main deps
to internal/ (so the future cmd/bintrail-console binary can reach them). This is
the leaf: deriveServerID has zero package-main dependencies (pure DSN-hash).

- deriveServerID → serverid.DeriveServerID (verbatim body; exported; serverid
  gains crypto/sha256, encoding/binary, go-sql-driver/mysql imports — none
  create a cycle: serverid is a leaf imported BY init/agent/up/monitor, imports
  none of them).
- Callers rewired: up.go (runUpStream), monitor.go (monitorSupervisor.Start).
- up.go drops the now-orphan crypto/sha256 + encoding/binary imports (mysql
  stays — used by other up.go paths).
- TestDeriveServerID moved to internal/serverid/serverid_test.go (deriveServerID
  → DeriveServerID; serverid_test gains fmt import).

Pure relocation, zero behavior change. Verified: go build/vet clean (unit +
integration); full unit suite green (incl. the moved DeriveServerID test); the
monitor/supervisor integration test passes against MySQL — it exercises
serverid.DeriveServerID end-to-end via monitorSupervisor.Start.
