+++ cmd/bintrail/up.go
@@ -2,9 +2,7 @@ package main
 
 import (
 	"context"
-	"crypto/sha256"
 	"database/sql"
-	"encoding/binary"
 	"fmt"
 	"log/slog"
 	"os"
@@ -19,6 +17,7 @@ import (
 	"github.com/dbtrail/bintrail/internal/config"
 	"github.com/dbtrail/bintrail/internal/console"
 	"github.com/dbtrail/bintrail/internal/indexer"
+	"github.com/dbtrail/bintrail/internal/serverid"
 	"github.com/dbtrail/bintrail/internal/streamrun"
 )
 
@@ -326,7 +325,7 @@ func runUpInit(cmd *cobra.Command) error {
 func runUpStream(cmd *cobra.Command, args []string) error {
 	serverID := upServerID
 	if serverID == 0 {
-		id, err := deriveServerID(upSourceDSN)
+		id, err := serverid.DeriveServerID(upSourceDSN)
 		if err != nil {
 			return fmt.Errorf("cannot auto-derive --server-id from --source-dsn: %w (pass --server-id explicitly to bypass)", err)
 		}
@@ -546,30 +545,4 @@ func populateStreamFlags(serverID uint32) {
 	strmReset = false
 	strmNoGapFill = false
 	strmGapTimeout = 30
-}
-
-// deriveServerID returns a deterministic uint32 server-id by hashing the
-// source DSN's host:user:dbname triple. The same DSN always produces the same
-// ID, so `bintrail up` resumes cleanly across restarts without the user
-// remembering what server-id they used last time.
-//
-// Returns an error when the DSN cannot be parsed — callers must handle this
-// rather than silently substituting a non-deterministic value, because a
-// per-invocation ID breaks the resume-from-checkpoint contract (MySQL would
-// treat each restart as a new replica).
-func deriveServerID(dsn string) (uint32, error) {
-	cfg, err := mysql.ParseDSN(dsn)
-	if err != nil {
-		return 0, fmt.Errorf("parse DSN: %w", err)
-	}
-	seed := fmt.Sprintf("%s|%s|%s", cfg.Addr, cfg.User, cfg.DBName)
-	sum := sha256.Sum256([]byte(seed))
-	raw := binary.BigEndian.Uint32(sum[:4])
-	// Map into [100000000, 4294967294]: subtract floor from uint32 range, mod
-	// into the resulting width, then add the floor back. Keeps the value high
-	// enough that collisions with typical hand-picked replica server-ids are
-	// unlikely.
-	const floor = uint32(100000000)
-	const width = uint32(4294967295 - floor) // 4194967295
-	return (raw % width) + floor, nil
 }
