+++ internal/serverid/serverid.go
@@ -6,10 +6,13 @@ package serverid
 
 import (
 	"context"
+	"crypto/sha256"
 	"database/sql"
+	"encoding/binary"
 	"errors"
 	"fmt"
 
+	"github.com/go-sql-driver/mysql"
 	"github.com/google/uuid"
 )
 
@@ -268,4 +271,30 @@ func DecommissionServer(ctx context.Context, db *sql.DB, bintrailID string) erro
 		return fmt.Errorf("bintrail_id %q not found or already decommissioned", bintrailID)
 	}
 	return nil
+}
+
+// DeriveServerID returns a deterministic uint32 server-id by hashing the
+// source DSN's host:user:dbname triple. The same DSN always produces the same
+// ID, so `bintrail up` resumes cleanly across restarts without the user
+// remembering what server-id they used last time.
+//
+// Returns an error when the DSN cannot be parsed — callers must handle this
+// rather than silently substituting a non-deterministic value, because a
+// per-invocation ID breaks the resume-from-checkpoint contract (MySQL would
+// treat each restart as a new replica).
+func DeriveServerID(dsn string) (uint32, error) {
+	cfg, err := mysql.ParseDSN(dsn)
+	if err != nil {
+		return 0, fmt.Errorf("parse DSN: %w", err)
+	}
+	seed := fmt.Sprintf("%s|%s|%s", cfg.Addr, cfg.User, cfg.DBName)
+	sum := sha256.Sum256([]byte(seed))
+	raw := binary.BigEndian.Uint32(sum[:4])
+	// Map into [100000000, 4294967294]: subtract floor from uint32 range, mod
+	// into the resulting width, then add the floor back. Keeps the value high
+	// enough that collisions with typical hand-picked replica server-ids are
+	// unlikely.
+	const floor = uint32(100000000)
+	const width = uint32(4294967295 - floor) // 4194967295
+	return (raw % width) + floor, nil
 }
